+++ app.py
@@ -1,47 +1,37 @@
-from flask import render_template, redirect, url_for, request, flash, send_from_directory, jsonify
-from flask_login import login_required, logout_user, login_user, current_user
-from werkzeug.security import generate_password_hash, check_password_hash
+from flask import render_template, redirect, url_for, send_from_directory
+from flask_login import login_required, current_user
 import config
-#from connexion.resolver import RelativeResolver
-from src.models import KalturaAppToken, AccessRestrictions, User, AppTokenSessionDefaults, \
-    UICustomizations, VendorProxies, db
+from src.models import User, UICustomizations, VendorProxies
 from config import login_manager
-import src.logger as logger, base64
-from src.auth_handler import get_user_credentials, swag_auth, generate_token
+from auth.routes import auth_bp
+from kaltura.routes import kaltura_bp
+from settings.routes import settings_bp
 
 @config.connex_app.app.context_processor
 def app_globals():
     app_title = UICustomizations.query.get(1)
     proxies = VendorProxies.query.get(1)
-    
     if app_title:
         title = app_title.integrator_title
     else:
-        title = 'Integration Manager'
-    
+        title = 'Integration Manager'  
     if proxies:
         kaltura_proxy_enabled = proxies.kaltura_proxy_enabled
         canvas_proxy_enabled = proxies.canvas_proxy_enabled
-
     return dict(custom_title=title,
                 kaltura_enabled=kaltura_proxy_enabled,
                 canvas_enabled=canvas_proxy_enabled)
 
 app = config.connex_app
 app.add_api(config.basedir / 'apispecs/swagger.yml', swagger_ui_options=config.swagoptions)
-
-#app.add_api(config.basedir / 'apispecs/swagger.yml', \
-#            swagger_ui_options=config.swagoptions, \
-#            resolver=RelativeResolver('apihandlers'))
-
-#app.add_api(config.basedir / 'apispecs/swagger-dev.yml', swagger_ui_options=config.swagoptions)
+app.app.register_blueprint(auth_bp, url_prefix='/auth')
+app.app.register_blueprint(kaltura_bp, url_prefix='/kaltura')
+app.app.register_blueprint(settings_bp, url_prefix='/settings')
 
 @login_manager.user_loader
 def user_loader(user_id):
     """Given *user_id*, return the associated User object.
-
     :param unicode user_id: user_id (email) user to retrieve
-
     """
     return User.query.get(user_id)
 
@@ -50,41 +40,7 @@ def home():
     if current_user.is_authenticated:
         return redirect(url_for('logpage'))
     else:
-        return redirect(url_for('login'))
-    
-@app.route('/main-config', methods=['GET', 'POST'])
-@login_required
-def main_config(): 
-    ui_customizations = UICustomizations.query.get(1)
-    vendor_proxies = VendorProxies.query.get(1)
-    
-    if request.method == 'GET':
-        return render_template('main-config.html', ui_customizations=ui_customizations,
-                               vendor_proxies=vendor_proxies)
-    
-    elif request.method == 'POST':
-        if request.args.get("f") == "ui":
-            ui_customizations.integrator_title = request.form.get('integrator-title')
-            db.session.add(ui_customizations)
-            db.session.commit()
-            flash('UI config updated', 'warning')
-            
-        elif request.args.get("f") == "vendorproxies":
-            vendor_proxies.kaltura_proxy_enabled = True if request.form.get('kaltura-proxy-enabled') else False
-            vendor_proxies.canvas_proxy_enabled = True if request.form.get('canvas-proxy-enabled') else False
-            db.session.add(vendor_proxies)
-            db.session.commit()
-            flash('Proxy configurations updated', 'warning')
-        
-        return redirect(url_for('main_config'))
-
-@app.route('/kaltura-manage')
-@login_required
-def kaltura_manage():
-    if VendorProxies.query.get(1).kaltura_proxy_enabled == False:
-        return redirect(url_for('main_config'))
-    tokens = KalturaAppToken.query.all()
-    return render_template("kaltura-manage.html", tokens=tokens)
+        return redirect(url_for('auth.login'))
 
 @app.route('/log')
 @login_required
@@ -97,44 +53,6 @@ def logpage():
 def send_report(path):
     return send_from_directory('logs', path)
 
-@app.route('/kaltura-config', methods=('GET', 'POST'))
-@login_required
-def kaltura_configpage():
-    if VendorProxies.query.get(1).kaltura_proxy_enabled == False:
-        return redirect(url_for('main_config'))
-    
-    access_restrictions = AccessRestrictions.query.get(1)
-    apptoken_session_defaults = AppTokenSessionDefaults.query.get(1)
-    
-    if request.method == 'GET':
-        return render_template('kaltura-config.html', access_restrictions=access_restrictions, \
-                               apptoken_session_defaults=apptoken_session_defaults)
-        
-    elif request.method == 'POST':
-        
-        if request.args.get("f") == "restrictions":
-            access_restrictions.allowed_categories = request.form.get('force-cats')
-            access_restrictions.force_labels = True if request.form.get('force-labels') else False
-            db.session.add(access_restrictions)
-            db.session.commit()
-            flash('Kaltura Access Restrictions config updated', 'warning')
-        elif request.args.get("f") == "sessions":
-            apptoken_session_defaults.partner_id = request.form.get('partner-id')
-            apptoken_session_defaults.session_expiry = request.form.get('session-expiry')
-            apptoken_session_defaults.use_local_storage = True if request.form.get('use-local-storage') else False
-            db.session.add(apptoken_session_defaults)
-            db.session.commit()
-            flash('Session Defaults config updated', 'warning')
-        
-        return redirect(url_for('kaltura_configpage'))
-    
-@app.route('/kaltura-addtokens')
-@login_required
-def addtokens():
-    if VendorProxies.query.get(1).kaltura_proxy_enabled == False:
-        return redirect(url_for('main_config'))
-    return render_template('kaltura-addtokens.html')
-
 @app.route('/apidocs')
 @login_required
 def apidocs():
@@ -144,90 +62,6 @@ def apidocs():
 @login_required
 def profile():
     return render_template('profile.html', username=current_user.username, role=current_user.role)
-
-@app.route('/login')
-def login():
-    return render_template('login.html')
-
-@app.route('/login', methods=['POST'])
-def login_post():
-
-    username = request.form.get('username')
-    password = request.form.get('password')
-    remember = True if request.form.get('remember') else False
-
-    user = User.query.filter_by(username=username).first()
-
-    # check if user actually exists
-    # take the user supplied password, hash it, and compare it to the hashed password in database
-    if not user or not check_password_hash(user.password, password):
-        flash('Please check your login details and try again.', 'warning')
-        logger.log("login attempt failed for user: " + username)
-        return redirect(url_for('login')) # if user doesn't exist or password is wrong, reload the page
-    else:
-        logger.log("login attempt succeeded for user: " + username)
-
-    # if the above check passes, then we know the user has the right credentials
-    login_user(user, remember=remember)
-    return redirect(url_for('logpage'))
-
-@app.route('/adduser', methods=['GET','POST'])
-@login_required
-def adduser():
-    if request.method == 'POST':
-        
-        # code to validate and add user to database goes here
-        username = request.form.get('username')
-        password = request.form.get('password')
-        email = request.form.get('email')
-        role = request.form.get('role')
-
-        user = User.query.filter_by(username=username).first() # if this returns a user, then the username already exists in database
-        emailcheck = User.query.filter_by(email=email).first()
-        if user:
-            flash('User already exists. Please try again', 'warning')
-            return redirect(url_for('adduser'))
-        
-        if emailcheck:
-            flash('Email already associated with another user. Please Try again', 'warning')
-            return redirect(url_for('adduser'))
-        
-        # create a new user with the form data. Hash the password so plaintext version isn't saved.
-        new_user = User(username=username, password=generate_password_hash(password, method='pbkdf2'), email=email, role=role)
-
-        # add the new user to the database
-        db.session.add(new_user)
-        db.session.commit()
-
-        flash('New user ' + username + ' added', 'success')
-        return redirect(url_for('adduser'))
-    
-    elif request.method == 'GET':
-        return render_template('adduser.html')
-        
-
-@app.route('/logout')
-@login_required
-def logout():
-    logout_user()
-    return redirect(url_for('login'))
-
-# route for internal authorizations needed by JS
-# performing requests. Admin auth needs is required
-# for additional session validation.
-@app.route('/get_auth_token')
-@login_required
-def get_auth_token():
-    username, password = get_user_credentials(username=current_user.username)
-    token = base64.b64encode(f"{username}:{password}".encode()).decode('utf-8')
-    return jsonify({'token': token})
-
-@app.route('/auth')
-@login_required
-def auth():
-    username, password = get_user_credentials(username=current_user.username)
-    jwt = generate_token(username)
-    return jsonify({'token': jwt})
     
 if __name__ == "__main__":
     app.run(host="0.0.0.0", port=8000)
