+++ templates/_note_create_form.html
@@ -12,7 +12,7 @@
   <div class="col-sm-10">
     <input class="form-control" name="content" type="text" value="">
   </div>
-  <button type="button" class="btn btn-info" data-action="create">
+  <button type="submit" class="btn btn-info" data-action="create">
     Create Note
   </button>
 </form>
