+++ templates/kaltura-config.html
@@ -4,7 +4,7 @@
 <div style="border-radius: 5px; padding: 20px">
   <h3>Kaltura Proxy Configuration</h3>
   <hr />
-  <form action="/kaltura-config?f=restrictions" method="post" id="restrictions">
+  <form action="/kaltura/config?f=restrictions" method="post" id="restrictions">
     <h4>Access Restrictions</h4>
     <label style="margin-top: 8px" for="force-cats"
       >Force Category Filtering (list allowed categories, separated by commas)</label
@@ -56,7 +56,7 @@
 </div>
 <hr />
 <div style="border-radius: 5px; padding: 20px">
-  <form action="/kaltura-config?f=sessions" method="post" id="sessions">
+  <form action="/kaltura/config?f=sessions" method="post" id="sessions">
     <h4>App Token Session Creation Defaults</h4>
     <label for="partner-id">Partner ID</label>
     <div class="col-sm-2">
