+++ templates/main-config.html
@@ -8,7 +8,7 @@
 <div style="border-radius: 5px; padding: 20px">
   <h3>Integrator Configuration</h3>
   <hr />
-  <form action="/main-config?f=ui" method="post" id="ui">
+  <form action="/settings/main-config?f=ui" method="post" id="ui">
     <h4>UI Customization</h4>
     <label for="integrator-title">Integrator Title</label>
     <div class="col-sm-3">
@@ -33,7 +33,7 @@
   </button>
   <hr />
   <h4>Vendor Proxies</h4>
-  <form action="/main-config?f=vendorproxies" method="post" id="vendorproxies">
+  <form action="/settings/main-config?f=vendorproxies" method="post" id="vendorproxies">
     <div class="form-check">
       <input
         class="form-check-input"
