+++ templates/profile.html
@@ -5,5 +5,5 @@
   Greetings, {{ username }}!
 </h3>
 <p>Role: {{ role }}</p>
-<a href="{{ url_for('auth') }}">Get a token</a>
+<a href="{{ url_for('auth.token') }}">Get a token</a>
 {% endblock %}
